moved db to /db/db.py moved formatters from parser added parseThisMonth method to parser
