+++ .gitignore
@@ -167,3 +167,6 @@ cython_debug/
 # Dbs
 *.csv
 *.csv.bak
+
+# Python scripts
+test.py
