+++ db/db.py
@@ -1,0 +1,7 @@
+import os
+
+from pymongo import MongoClient
+
+client = MongoClient(os.environ.get('MONGODB_ADRESS'), connect=True)
+db = client.ruzbotdb
+users = db.users
