+++ ruzparser/ruzparser.py
@@ -1,3 +1,4 @@
+import calendar
 import json
 from datetime import datetime, timedelta
 import logging
@@ -5,27 +6,9 @@ import logging
 import aiohttp
 import requests
 
-lession_number = {
-    "08:30": 1,
-    "10:10": 2,
-    "12:40": 3,
-    "14:20": 4,
-    "16:00": 5,
-    "17:40": 6
-}
-
-week_days = {
-    0: "Понедельник",
-    1: "Вторник",
-    2: "Среда",
-    3: "Четверг",
-    4: "Пятница",
-    5: "Суббота",
-    6: "Воскресенье"
-}
-
 GROUP_URL = "https://ruz.mstuca.ru/api/search?term={}&type=group"
 LESSIONS_URL = "https://ruz.mstuca.ru/api/schedule/group/{}?start={}&finish={}&lng=1"
+
 
 class RuzParser:
     """
@@ -116,79 +99,29 @@ class RuzParser:
         
         logging.info(f'parseWeek: {start} {end}')
         return await self.parse(group, start, end)
-            
-    def formatDay(self, data, _timedelta = 0):
-        """
-        Format schedule for group for one day
-        
-        Args:
-            data (dict): Schedule in JSON format
-            _timedelta (int): Timedelta in days. Default is 0
-        
-        Returns:
-            str: Formatted schedule
-        """
-        date = datetime.today() + timedelta(days = _timedelta)
-        week_day = week_days.get(date.weekday())
-        date = date.strftime('%d.%m')
-        
-        lessions = ""
-        for i in range(len(data)): 
-            lessions += f"-- {lession_number.get(data[i].get('beginLesson'))} пара [{data[i].get('beginLesson')} - {data[i].get('endLesson')}] --" + '\n'
-            lessions += data[i].get("discipline") + f" ({self.parseKindOfWork(data[i].get('kindOfWork'))})" + '\n'
-            lessions += f"Аудитория: {data[i].get('auditorium').split('/')[1]}" + '\n'
-            lessions += f"Преподаватель: {data[i].get('lecturer_title')}, {data[i].get('lecturer_rank')}" + '\n'
-        
-        if lessions == "":
-            return f"= {date} = \n\nПар нет"
-        
-        return f"= {week_day} ({date}) = \n{lessions}"
     
-    def formatWeek(self, data):
+    async def parseThisMonth(self, group):
         """
-        Format schedule for group for one week
+        Parse schedule for group for one week
         
         Args:
-            data (dict): Schedule in JSON format
+            group (str): Group name
+            _timedelta (int): Timedelta in weeks. Default is 0
         
         Returns:
-            str: Formatted schedule
+            dict: Schedule in JSON format
         """
-        dates = {
-        }
-        for i in range(len(data)):
-            datetime_object = datetime.strptime(data[i].get("date"), '%Y-%m-%d')
-            week_day = week_days.get(datetime_object.weekday())
-            dates.update(
-                {data[i].get("date"): f"_= {week_day} ({'.'.join(data[i].get('date').split('-')[1:])}) =_ \n"}
-                )
+        logging.info(f'parsing this month for group {group}')
         
-        for i in range(len(data)):
-            tmp = dates.get(data[i].get("date"))
-            tmp += f"*-- {lession_number.get(data[i].get('beginLesson'))} пара [{data[i].get('beginLesson')} - {data[i].get('endLesson')}] --*" + '\n  '
-            tmp += data[i].get("discipline") + f" ({self.parseKindOfWork(data[i].get('kindOfWork'))})" + '\n  '
-            tmp += f"Аудитория: {data[i].get('auditorium').split('/')[1]}" + '\n  '
-            tmp += f"Преподаватель: {data[i].get('lecturer_title')}, {data[i].get('lecturer_rank')}" + '\n'
-            
-            dates.update({data[i].get("date"): tmp})
-        
-        # split this by "-": list(dates.keys())[0] and replace "-" with dost, then conver to string
-        
-        if len(list(dates.keys())) == 0:
-            return "Пар нет"
-        
-        lessions = "== Расписание на неделю {} - {} == \n\n".format(
-            ".".join(list(dates.keys())[0].split('-'))[5:], 
-            ".".join(list(dates.keys())[-1].split('-'))[5:]
+        first_day_of_month = datetime.today().replace(day=1)
+        last_day_of_month =first_day_of_month + timedelta(
+            days=calendar.monthrange(first_day_of_month.year, first_day_of_month.month)[1] - 1
             )
-        for key in dates.keys():
-            lessions += dates.get(key) + "\n"
         
-        replacables = ['.', '-', '(', ')', "="]
-        for ch in replacables:            
-            lessions = lessions.replace(ch, f"\\{ch}")
+        start = first_day_of_month.strftime('%Y.%m.%d')
+        end = last_day_of_month.strftime('%Y.%m.%d')
         
-        return lessions
+        return await self.parse(group, start, end)
     
     async def search_group(self, group_name):
         """
@@ -198,15 +131,3 @@ class RuzParser:
             json = await self.fetch(session, GROUP_URL.format(group_name))
         
         return json
-    
-    @staticmethod
-    def parseKindOfWork(kindOfWork):
-        """
-        Parse kind of work
-        """
-        if kindOfWork == "Лекция":
-            return "Лек."
-        if kindOfWork == "Практические (семинарские) занятия":
-            return "Пр. зан."
-        
-        return kindOfWork
