+++ utils/__init__.py
@@ -1,3 +1,3 @@
-
 from .user import Users
 from .utils import RANDOM_GROUP_NAMES
+from .formatters import *
